requirements.txt
# BioValue-AI Python Dependencies
# LLM Bridge Service

# gRPC
grpcio>=1.62.0
grpcio-tools>=1.62.0
protobuf>=4.25.0

# LLM Providers
openai>=1.12.0
anthropic>=0.18.0

# MCP Client
mcp>=0.9.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.1

# Async Support
asyncio>=3.4.3
aiohttp>=3.9.3

# Vector DB
pymilvus>=2.4.0

# Observability
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp>=1.24.0
prometheus-client>=0.20.0

# Redis
redis>=5.0.0
hiredis>=2.3.0

# Utilities
pydantic>=2.6.0
structlog>=24.1.0
tenacity>=8.2.0

